Allow to specify user_data in a template
For some cases relying on a script to be passed to shell
provisioner is not enough. So let's allow user to specify
actual userdata to be passed to the vm on creation, this
will help in complex scenarios like persistently changing
network routing that will survive reboots.

Sample configuration:

"user_data": "#cloud-config\nbootcmd:\n- ip route add 10.0.0.0/8 via 10.130.68.1 dev eth0"